+++ src/components/accommodation/contents/address/AccommodationAddressCheckContents.tsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
-import ContentsTitle from './ContentsTitle';
-import ContentsSubText from './ContentsSubText';
+import ContentsTitle from '../ContentsTitle';
+import ContentsSubText from '../ContentsSubText';
 import AddressInputContents from './AddressInputContents';
 
 export default function AccommodationAddressCheckContents() {
