+++ src/components/accommodation/contents/address/AccommodationAddressContents.tsx
@@ -1,7 +1,7 @@
 import styled from 'styled-components';
 import AddressSearchBar from './AddressSearchBar';
-import ContentsTitle from './ContentsTitle';
-import ContentsSubText from './ContentsSubText';
+import ContentsTitle from '../ContentsTitle';
+import ContentsSubText from '../ContentsSubText';
 
 export default function AccommodationAddressContents() {
   return (
