+++ src/components/accommodation/contents/notUsed/AccommodationCountInfo.tsx
