+++ src/components/accommodation/contents/notUsed/AccommodationFeeContents.tsx
@@ -1,5 +1,5 @@
-import ContentsTitle from './ContentsTitle';
-import ContentsSubText from './ContentsSubText';
+import ContentsTitle from '../ContentsTitle';
+import ContentsSubText from '../ContentsSubText';
 import AccommodationSettingFeeBox from './AccommodationSettingFeeBox';
 
 export default function AccommodationFeeContents() {
