+++ src/components/accommodation/contents/notUsed/RoomTypeItem.tsx
@@ -1,22 +1,24 @@
 import styled from 'styled-components';
-import { AccommodationIconMap, AccommodationNameMap, AccommodationType } from '../../../constants/accommodationType';
+import { RoomIconMap, RoomNameMap, RoomContentMap, RoomType } from '../../../../constants/roomType';
 
-interface AccommodationTypeItemProps {
-  type: keyof AccommodationType;
+interface RoomTypeItemProps {
+  type: keyof RoomType;
   isSelected: boolean;
-  onClick: (type: keyof AccommodationType) => void;
+  onClick: (type: keyof RoomType) => void;
 }
 
-export default function AccommodationTypeItem({ type, isSelected, onClick }: AccommodationTypeItemProps) {
+export default function RoomTypeItem({ type, isSelected, onClick }: RoomTypeItemProps) {
   const handleClick = () => {
     onClick(type);
   };
-
   return (
     <StyledButtonDiv>
       <StyledItemButton value={type} type="button" role="checkbox" aria-checked={isSelected} onClick={handleClick}>
-        {AccommodationIconMap[type]}
-        <StyledItemName>{AccommodationNameMap[type]}</StyledItemName>
+        <StyledTextContainer>
+          <StyledItemName>{RoomNameMap[type]}</StyledItemName>
+          <StyledStepContent>{RoomContentMap[type]}</StyledStepContent>
+        </StyledTextContainer>
+        {RoomIconMap[type]}
       </StyledItemButton>
     </StyledButtonDiv>
   );
@@ -28,13 +30,13 @@ const StyledButtonDiv = styled.div`
 
 const StyledItemButton = styled.button`
   display: flex;
-  flex-direction: column;
-  justify-content: center;
-  width: 200px;
-  height: 110px;
+  justify-content: space-between;
+  align-items: center;
+  width: 550px;
+  height: 90px;
   border: 2px solid rgba(0, 0, 0, 0.2);
   border-radius: 10px;
-  padding-left: 10px;
+  padding: 0 20px;
   background-color: white;
 
   &:hover {
@@ -57,7 +59,20 @@ const StyledItemButton = styled.button`
   }}
 `;
 
+const StyledTextContainer = styled.div`
+  width: 400px;
+  display: flex;
+  flex-direction: column;
+`;
+
 const StyledItemName = styled.span`
   font-size: 15px;
-  margin: 10px 0 0 10px;
+  font-weight: 500;
+  text-align: left;
+`;
+
+const StyledStepContent = styled.p`
+  font-size: 13px;
+  color: rgba(0, 0, 0, 0.4);
+  text-align: left;
 `;
