+++ src/components/accommodation/contents/text/AccommodationDetailContents.tsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
-import ContentsTitle from './ContentsTitle';
-import ContentsSubText from './ContentsSubText';
+import ContentsTitle from '../ContentsTitle';
+import ContentsSubText from '../ContentsSubText';
 import AccommodationTextBox from './AccommodationTextBox';
 
 export default function AccommodationDetailContents() {
